Split protocol
