+++ src/protocol/response.rs
@@ -1,41 +1,6 @@
-//! The main protocol handling
+//! Response messages
 
-#[cfg(feature = "backend")]
-use schema::sessions;
-use serde_cbor::to_vec;
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-/// All possible request variants
-pub enum Request {
-    /// Possible variants of a login request
-    Login(Login),
-
-    /// A logout request with a provided session
-    Logout(Session),
-}
-
-impl Request {
-    /// Consume the object into a vector
-    pub fn to_vec(&self) -> Option<Vec<u8>> {
-        to_vec(self).ok()
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-/// Possible login request variants
-pub enum Login {
-    /// A credentials based request
-    Credentials {
-        /// The username
-        username: String,
-
-        /// The password
-        password: String,
-    },
-
-    /// A session based request
-    Session(Session),
-}
+use protocol::model::Session;
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
 /// All possible response variants
@@ -83,13 +48,4 @@ pub enum ResponseError {
     #[fail(display = "unable to delete session within database")]
     /// Session deletion in database failed
     DeleteSession,
-}
-
-#[cfg_attr(feature = "backend", derive(Insertable, Queryable))]
-#[cfg_attr(feature = "backend", table_name = "sessions")]
-#[derive(Clone, Debug, Deserialize, Serialize)]
-/// A session representation
-pub struct Session {
-    /// The actual session token
-    pub token: String,
 }
